double return for step(), another copy to reference

+++ stan/math/prim/scal/fun/step.hpp
@@ -25,7 +25,7 @@ namespace stan {
      * @return zero if the value is less than zero, and one otherwise
      */
     template <typename T>
-    inline int step(const T& y) {
+    inline double step(const T& y) {
       return y < 0.0 ? 0 : 1;
     }
